Move all the converters to their own files.
This PR moves all the converters to their own files.

Future work should similarly split up the tests.
